requirements.txt
google-generativeai
streamlit
python-dotenv
sajupy
